+++ android/src/main/kotlin/me/leoletto/caller/CallerPhoneStateListener.kt
@@ -1,6 +1,7 @@
 package me.leoletto.caller
 
 import android.content.Context
+import android.content.Intent
 import android.os.Build
 import android.telephony.PhoneStateListener
 import android.telephony.TelephonyManager
@@ -15,27 +16,75 @@ import java.time.Duration
 import java.time.ZonedDateTime
 import java.util.ArrayList
 
-class CallerPhoneStateListener internal constructor(private val context: Context, private val flutterLoader: FlutterLoader) : PhoneStateListener() {
-    private var previousState = 0
+enum class CallType {
+    INCOMING, OUTGOING;
+}
+
+class CallerPhoneStateListener internal constructor(
+    private val context: Context,
+    private val intent: Intent,
+    private val flutterLoader: FlutterLoader
+) : PhoneStateListener() {
+
     private var sBackgroundFlutterEngine: FlutterEngine? = null
     private var channel: MethodChannel? = null
     private var callbackHandler: Long? = null
     private var callbackHandlerUser: Long? = null
+
     private var time: ZonedDateTime? = null
+    private var callType: CallType? = null
+    private var previousState: Int? = null
 
     @RequiresApi(Build.VERSION_CODES.O)
     @Synchronized
-    override fun onCallStateChanged(state: Int, incomingNumber: String) {
-        // Prepare the arguments to be sent to the callback
+    override fun onCallStateChanged(state: Int, incomingNumber: String?) {
+        when (state) {
+            TelephonyManager.CALL_STATE_IDLE -> {
+                val duration = Duration.between(time ?: ZonedDateTime.now(), ZonedDateTime.now())
+
+                if (previousState == TelephonyManager.CALL_STATE_OFFHOOK && callType == CallType.INCOMING) {
+                    // Incoming call ended
+                    Log.d(CallerPlugin.PLUGIN_NAME, "Phone State event IDLE (INCOMING ENDED) with number - $incomingNumber")
+                    notifyFlutterEngine(CallType.INCOMING, duration.toMillis() / 1000, incomingNumber!!)
+                } else if(callType == CallType.OUTGOING) {
+                    // Outgoing call ended
+                    Log.d(CallerPlugin.PLUGIN_NAME, "Phone State event IDLE (OUTGOING ENDED) with number - $incomingNumber")
+                    notifyFlutterEngine(CallType.OUTGOING, duration.toMillis() / 1000, incomingNumber!!)
+                }
+
+                callType = null
+                previousState = TelephonyManager.CALL_STATE_IDLE
+            }
+            TelephonyManager.CALL_STATE_OFFHOOK -> {
+                Log.d(CallerPlugin.PLUGIN_NAME, "Phone State event STATE_OFFHOOK")
+                // Phone didn't ring, so this is an outgoing call
+                if (callType == null)
+                    callType = CallType.OUTGOING
+
+                // Get current time to use later to calculate the duration of the call
+                time = ZonedDateTime.now()
+                previousState = TelephonyManager.CALL_STATE_OFFHOOK
+            }
+            TelephonyManager.CALL_STATE_RINGING -> {
+                Log.d(CallerPlugin.PLUGIN_NAME, "Phone State event PHONE_RINGING number: $incomingNumber")
+                callType = CallType.INCOMING
+                previousState = TelephonyManager.CALL_STATE_RINGING
+            }
+        }
+    }
+
+    private fun notifyFlutterEngine(type: CallType, duration: Long, number: String){
         val arguments = ArrayList<Any?>()
+
+        // Initialize flutter engine
         if (sBackgroundFlutterEngine == null) {
             callbackHandler = context.getSharedPreferences(
-                    CallerPlugin.PLUGIN_NAME,
-                    Context.MODE_PRIVATE
+                CallerPlugin.PLUGIN_NAME,
+                Context.MODE_PRIVATE
             ).getLong(CallerPlugin.CALLBACK_SHAREDPREFERENCES_KEY, 0)
             callbackHandlerUser = context.getSharedPreferences(
-                    CallerPlugin.PLUGIN_NAME,
-                    Context.MODE_PRIVATE
+                CallerPlugin.PLUGIN_NAME,
+                Context.MODE_PRIVATE
             ).getLong(CallerPlugin.CALLBACK_USER_SHAREDPREFERENCES_KEY, 0)
             if (callbackHandler == 0L || callbackHandlerUser == 0L) {
                 Log.e(CallerPlugin.PLUGIN_NAME, "Fatal: No callback registered")
@@ -52,63 +101,26 @@ class CallerPhoneStateListener internal constructor(private val context: Context
             }
             sBackgroundFlutterEngine = FlutterEngine(context)
             val args = DartCallback(
-                    context.assets,
-                    flutterLoader.findAppBundlePath(),
-                    callbackInfo
+                context.assets,
+                flutterLoader.findAppBundlePath(),
+                callbackInfo
             )
 
             // Start running callback dispatcher code in our background FlutterEngine instance.
             sBackgroundFlutterEngine!!.dartExecutor.executeDartCallback(args)
         }
+        // Create the MethodChannel used to communicate between the callback
+        // dispatcher and this instance.
+        channel = MethodChannel(
+            sBackgroundFlutterEngine!!.dartExecutor.binaryMessenger,
+            CallerPlugin.PLUGIN_NAME + "_background"
+        )
+
         arguments.add(callbackHandler)
         arguments.add(callbackHandlerUser)
-        arguments.add(incomingNumber)
-
-        // Create the MethodChannel used to communicate between the callback
-        // dispatcher and this GeofencingService instance.
-        channel = MethodChannel(
-                sBackgroundFlutterEngine!!.dartExecutor.binaryMessenger,
-                CallerPlugin.PLUGIN_NAME + "_background"
-        )
-        when (state) {
-            TelephonyManager.CALL_STATE_IDLE -> {
-                if (previousState == TelephonyManager.CALL_STATE_OFFHOOK) {
-                    val now = ZonedDateTime.now()
-                    val duration = Duration.between(time, now)
-                    Log.d(CallerPlugin.PLUGIN_NAME, "Phone State event CALL_ENDED with number $incomingNumber")
-                    arguments.add("callEnded")
-                    arguments.add(duration.toMillis() / 1000)
-                    time = null
-                    channel!!.invokeMethod("incomingCallEnded", arguments)
-                } else if (previousState == TelephonyManager.CALL_STATE_RINGING) {
-                    Log.d(CallerPlugin.PLUGIN_NAME, "Phone State event INCOMING_CALL_MISSED $incomingNumber")
-                    arguments.add("onMissedCall")
-                    arguments.add(null)
-                    channel!!.invokeMethod("onMissedCall", arguments)
-                }
-                previousState = TelephonyManager.CALL_STATE_IDLE
-            }
-            TelephonyManager.CALL_STATE_OFFHOOK -> {
-                if (previousState == TelephonyManager.CALL_STATE_RINGING) {
-                    Log.d(CallerPlugin.PLUGIN_NAME, "Phone State event INCOMING_CALL_ANSWERED $incomingNumber")
-                    arguments.add("onIncomingCallAnswered")
-                    arguments.add(null)
-                    channel!!.invokeMethod("onIncomingCallAnswered", arguments)
-                }
-                previousState = TelephonyManager.CALL_STATE_OFFHOOK
-
-                // Get the current time to count later the duration of the call
-                time = ZonedDateTime.now()
-            }
-            TelephonyManager.CALL_STATE_RINGING -> {
-                if (previousState == TelephonyManager.CALL_STATE_IDLE) {
-                    Log.d(CallerPlugin.PLUGIN_NAME, "Phone State event INCOMING_CALL_RECEIVED $incomingNumber")
-                    arguments.add("onIncomingCallReceived")
-                    arguments.add(null)
-                    channel!!.invokeMethod("onIncomingCallReceived", arguments)
-                }
-                previousState = TelephonyManager.CALL_STATE_RINGING
-            }
-        }
+        arguments.add(type.toString())
+        arguments.add(duration)
+        arguments.add(number)
+        channel!!.invokeMethod("call", arguments)
     }
 }