+++ example/lib/main.dart
@@ -2,47 +2,26 @@ import 'package:caller/caller.dart';
 import 'package:flutter/material.dart';
 
 /// Defines a callback that will handle all background incoming events
-///
-/// The duration will only have a value if the current event is `CallerEvent.callEnded`
 Future<void> callerCallbackHandler(
   CallerEvent event,
   String number,
-  int? duration,
+  int duration,
 ) async {
   print("New event received from native $event");
   switch (event) {
-    case CallerEvent.callEnded:
+    case CallerEvent.incoming:
       print(
-          '[ Caller ] Ended a call with number $number and duration $duration');
+          '[ Caller ] Incoming call ended, number: $number, duration $duration s');
       break;
-    case CallerEvent.onMissedCall:
-      print('[ Caller ] Missed a call from number $number');
-      break;
-    case CallerEvent.onIncomingCallAnswered:
-      print('[ Caller ] Accepted call from number $number');
-      break;
-    case CallerEvent.onIncomingCallReceived:
-      print('[ Caller ] Phone is ringing with number $number');
+    case CallerEvent.outgoing:
+      print(
+          '[ Caller ] Ougoing call ended, number: $number, duration: $duration s');
       break;
   }
 }
 
-Future<void> initialize() async {
-  /// Check if the user has granted permissions
-  final permission = await Caller.checkPermission();
-  print('Caller permission $permission');
-
-  /// If not, then request user permission to access the Call State
-  if (!permission)
-    Caller.requestPermissions();
-  else
-    Caller.initialize(callerCallbackHandler);
-}
-
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
-
-  initialize();
 
   /// Run your app as you would normally do...
   runApp(MyApp());
@@ -55,36 +34,17 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
+      title: 'Caller Plugin Demo',
       theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
         primarySwatch: Colors.blue,
       ),
-      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+      home: const MyHomePage(title: 'Caller Plugin Demo'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({Key? key, required this.title}) : super(key: key);
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
-
   final String title;
 
   @override
@@ -92,68 +52,61 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
+  bool? hasPermission;
 
-  void _incrementCounter() {
-    setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun the build method below
-      // so that the display can reflect the updated values. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
-      _counter++;
-    });
+  @override
+  void initState() {
+    super.initState();
+
+    _checkPermission();
+  }
+
+  Future<void> _checkPermission() async {
+    final permission = await Caller.checkPermission();
+    print('Caller permission $permission');
+    setState(() => hasPermission = permission);
+  }
+
+  Future<void> _requestPermission() async {
+    await Caller.requestPermissions();
+    await _checkPermission();
+  }
+
+  Future<void> _stopCaller() async {
+    await Caller.stopCaller();
+  }
+
+  Future<void> _startCaller() async {
+    if (hasPermission != true) return;
+    await Caller.initialize(callerCallbackHandler);
   }
 
   @override
   Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
     return Scaffold(
       appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
         title: Text(widget.title),
       ),
       body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
         child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Invoke "debug painting" (press "p" in the console, choose the
-          // "Toggle Debug Paint" action from the Flutter Inspector in Android
-          // Studio, or the "Toggle Debug Paint" command in Visual Studio Code)
-          // to see the wireframe for each widget.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            const Text(
-              'You have pushed the button this many times:',
+            Text(hasPermission == true ? 'Has permission' : 'No permission'),
+            ElevatedButton(
+              onPressed: () => _requestPermission(),
+              child: Text('Ask Permission'),
             ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headline4,
+            ElevatedButton(
+              onPressed: () => _startCaller(),
+              child: Text('Start caller'),
+            ),
+            ElevatedButton(
+              onPressed: () => _stopCaller(),
+              child: Text('Stop caller'),
             ),
           ],
         ),
       ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
     );
   }
 }