+++ lib/src/caller_event.dart
@@ -1,6 +1,1 @@
-enum CallerEvent {
-  callEnded,
-  onMissedCall,
-  onIncomingCallAnswered,
-  onIncomingCallReceived,
-}
+enum CallerEvent { incoming, outgoing }