+++ lib/src/caller_handler.dart
@@ -1,8 +1,0 @@
-abstract class CallerHandler {
-  void onIncomingCallReceived();
-  void onIncomingCallAnswered();
-  void onIncomingCallEnded();
-  void onOutgoingCallStarted();
-  void onOutgoingCallEnded();
-  void onMissedCall();
-}